Parcial 2/envios/requirements.txt
ariadne==0.23.0
Flask==3.0.3
SQLAlchemy==2.0.35
PyMySQL==1.1.1
grpcio==1.66.1
